[core] Introduce RenderOrchestrator
