+++ src/core-files.json
@@ -127,6 +127,7 @@
         "src/mbgl/renderer/pattern_atlas.cpp",
         "src/mbgl/renderer/render_layer.cpp",
         "src/mbgl/renderer/render_light.cpp",
+        "src/mbgl/renderer/render_orchestrator.cpp",
         "src/mbgl/renderer/render_source.cpp",
         "src/mbgl/renderer/render_static_data.cpp",
         "src/mbgl/renderer/render_tile.cpp",
@@ -643,6 +644,7 @@
         "mbgl/renderer/property_evaluator.hpp": "src/mbgl/renderer/property_evaluator.hpp",
         "mbgl/renderer/render_layer.hpp": "src/mbgl/renderer/render_layer.hpp",
         "mbgl/renderer/render_light.hpp": "src/mbgl/renderer/render_light.hpp",
+        "mbgl/renderer/render_orchestrator.hpp": "src/mbgl/renderer/render_orchestrator.hpp",
         "mbgl/renderer/render_pass.hpp": "src/mbgl/renderer/render_pass.hpp",
         "mbgl/renderer/render_source.hpp": "src/mbgl/renderer/render_source.hpp",
         "mbgl/renderer/render_source_observer.hpp": "src/mbgl/renderer/render_source_observer.hpp",
